participants api

+++ backend/src/main/java/com/xjc/harlab/repository/AttacksRepository.java
@@ -1,11 +1,9 @@
 package com.xjc.harlab.repository;
 
-import com.xjc.harlab.model.Attacks;
+import com.xjc.harlab.model.entity.Attacks;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import java.math.BigDecimal;
-
-public interface AttacksRepository extends JpaRepository<Attacks, BigDecimal>, JpaSpecificationExecutor<Attacks> {
+public interface AttacksRepository extends JpaRepository<Attacks, Integer>, JpaSpecificationExecutor<Attacks> {
 
 }
